working on doctor chat with swipe
