+++ client/src/Components/Globals/Apps/DoctorApp.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { BrowserRouter, Route, Switch, Redirect } from "react-router-dom";
-import DoctorContextProvider from '../../../Contexts/Doctor.context';
-import DoctorChat from "../../Pages/DoctorChat/DoctorChat";
+import DoctorContextProvider from "../../../Contexts/Doctor.context";
+import DoctorChat2 from "../../Pages/DoctorChat/DoctorChat2";
 import PatientQueue from "../../Pages/PatientQueue/PatientQueue";
 
 const DoctorApp = () => {
@@ -10,7 +10,7 @@ const DoctorApp = () => {
       <BrowserRouter>
         <Switch>
           <Route exact path="/doctor/consultation">
-            <DoctorChat />
+            <DoctorChat2 />
           </Route>
           <Route exact path="/doctor/queue">
             <PatientQueue />
@@ -19,7 +19,6 @@ const DoctorApp = () => {
         </Switch>
       </BrowserRouter>
     </DoctorContextProvider>
-
   );
 };
